+++ lib/Applications.ps1
@@ -3,7 +3,7 @@
     @('Helpers', 'New-IssuePrompt'),
     @('json', 'ConvertToPrettyJson'),
     @('manifest', 'Resolve-ManifestInformation'),
-    @('depends', 'script_deps'),
+    @('Dependencies', 'Resolve-DependsProperty'),
     @('Versions', 'Clear-InstalledVersion')
 ) | ForEach-Object {
     if (!([bool] (Get-Command $_[1] -ErrorAction 'Ignore'))) {
@@ -192,8 +192,8 @@ function app_status($app, $global) {
     }
 
     $status.missing_deps = @()
-    # TODO: Adopt Resolve-ManifestInformation
-    $deps = @(runtime_deps $manifest) | Where-Object {
+    # TODO: Adopt Resolve-ManifestInformation not needed to be fully compatible, consider some simple parsing
+    $deps = @(Resolve-DependsProperty -Manifest $manifest) | Where-Object {
         $app, $bucket, $null = parse_app $_
         return !(installed $app)
     }