+++ libexec/scoop-status.ps1
@@ -17,6 +17,7 @@
     @('Helpers', 'New-IssuePrompt'),
     @('Applications', 'Get-InstalledApplicationInformation'),
     @('buckets', 'Get-KnownBucket'),
+    @('Dependencies', 'Resolve-DependsProperty'),
     @('depends', 'script_deps'),
     @('Git', 'Invoke-GitCmd'),
     @('manifest', 'Resolve-ManifestInformation'),